+++ packages/psss/src/constants/index.js
@@ -4,4 +4,5 @@
  */
 
 export * from './colors';
+export * from './syndromes';
 export * from './weeklyReports';
