+++ packages/psss/src/constants/syndromes.js
@@ -1,0 +1,12 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+export const SYNDROMES = {
+  AFR: 'Acute Fever and Rash (AFR)',
+  DIA: 'Diarrhoea (DIA)',
+  ILI: 'Influenza-like Illness (ILI)',
+  PF: 'Prolonged Fever (PF)',
+  DLI: 'Dengue-like Illness (DLI)',
+};
