+++ packages/psss/src/constants/weeklyReports.js
@@ -21,11 +21,3 @@ export const TABLE_STATUSES = {
   SAVING: 'saving',
   ERROR: 'error',
 };
-
-export const SYNDROMES = {
-  AFR: 'Acute Fever and Rash (AFR)',
-  DIA: 'Diarrhoea (DIA)',
-  ILI: 'Influenza-like Illness (ILI)',
-  PF: 'Prolonged Fever (PF)',
-  DLI: 'Dengue-like Illness (DLI)',
-};
