+++ packages/psss/src/containers/Tables/OutbreaksTable.js
@@ -82,5 +82,5 @@ OutbreaksTable.propTypes = {
 };
 
 OutbreaksTable.defaultProps = {
-  countryCode: null,
+  countryCode: '',
 };
